Fix input bug
Previously, closing the window and pressing buttons would result in lag
that would be several seconds long. The reason for this is because SDL
keeps an internal event queue that can only be cleared by
SDL_PollEvent(). As I only called the function a single time per frame,
events accumulated very quickly whenever you moved the mouse on screen
causing the window closing event to be pushed back until mouse events were
fully processed.

The fix for this was to put SDL_PollEvent() in a while loop until all
events were handled for that frame. This makes sure that the window
closed event and mouse click events are processed on the next frame they
were given, solving the issue.