Fix fatal segmentation fault
Usually, the game would have a random chance of a segfault when you
died. After debugging, I found out Snek's current_pieces member was
never formally initialized -- just added to and read from. This lead to
undefined behaviour every time a Snek object was dynamically allocated
again.

For some reason, this didn't cause the game to cause immediantly, and
not at all in Windows. The reason for this is probably the compiler: the
compiler for windows probably initializes all variables for you, while
g++ and clang don't.

The actual reason for crashing is still unknown. Since current_pieces is
directly attached to the scoreboard, I would've expected the score value
to wildly jump values as different pieces of garbage were read from
memory, though that is not the case. Weirdly enough, manually inputting
large values in the hundreads of thousands in current_pieces also led to
a crash, leading me to believe the code responsible for displaying the
score can't handle super large values.