Add installSDL.bash for easy installation
To install all of the modules of SDL2 correctly and without hassle,
create a bash script that will do the work for you. In addition to this,
move the executable into executable/ to prevent the root from being
messy.

As we have this script to install everything, also remove all .dll that
came with the project